tools/python/deep-fusion-landcover/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "deep-fusion-landcover"
version = "1.0.0"
description = "Deep-fusion multi-sensor landcover and landcover-change mapper for Austin, TX — fusing Landsat 4/5/7/8/9, Sentinel-1/2, NAIP, 3DEP LiDAR, GEDI, and NLCD across 35 years with pseudo-quantum VQC, TorchGeo CNN (ONNX), SAM2 segmentation, and a LightGBM meta-stacker ensemble"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
    # STAC / Planetary Computer
    "pystac-client>=0.7",
    "planetary-computer>=1.0",
    "stackstac>=0.5",
    "pystac>=1.10",

    # Raster I/O and array processing
    "rioxarray>=0.15",
    "xarray>=2024.1",
    "numpy>=1.26",
    "dask[array]>=2024.1",
    "rasterio>=1.3",

    # Vector I/O
    "geopandas>=0.14",
    "pyogrio>=0.7",
    "shapely>=2.0",

    # Image processing / classical ML
    "scikit-learn>=1.4",
    "scipy>=1.12",
    "scikit-image>=0.22",

    # Gradient boosting meta-learner
    "lightgbm>=4.3",

    # Deep learning / TorchGeo CNN encoder
    "torch>=2.2",
    "torchvision>=0.17",
    "torchgeo>=0.5",

    # ONNX export + CPU inference
    "onnx>=1.16",
    "onnxruntime>=1.18",

    # LiDAR point cloud processing
    "laspy[lazrs]>=2.5",

    # Terrain / hydrology
    "richdem>=0.3",

    # CRS
    "pyproj>=3.6",

    # Visualization
    "matplotlib>=3.8",
    "folium>=0.16",
    "branca>=0.7",
    "Pillow>=10.0",

    # HTTP (NLCD, EarthData downloads)
    "requests>=2.31",

    # CLI
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]
sam = [
    # SAM2 must be installed from Meta's repo or via pip if published:
    # pip install segment-anything-2
    "segment-anything-2>=1.0",
]
earthdata = [
    # For GEDI LP DAAC access
    "earthaccess>=0.9",
]

[project.scripts]
deep-fusion-landcover = "deep_fusion_landcover.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/deep_fusion_landcover"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "D", "UP", "B", "SIM"]
ignore = ["D203", "D213"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --tb=short"
